Refactor backend code into directories, adding init pys cause fastapi kept breaking might break deployment
