+++ backend/States/IOSchema.py
@@ -5,7 +5,7 @@ from typing import List, Optional, Literal, Annotated
 from fastapi import UploadFile, Form
 from pydantic import BaseModel
 
-from Enums import Roles
+from backend.States.Enums import Roles
 
 
 class InviteInput(BaseModel):
